main/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
dj-database-url==3.1.2
Django==6.0.2
django-allauth==65.14.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.2.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
pycparser==3.0
PyJWT==2.11.0
python-decouple==3.8
python-openid==2.2.5
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
steamid-converter==1.0.1
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
django-cors-headers
whitenoise
gunicorn
dj-database-url
psycopg2-binary
python-decouple
django-filter
social-auth-app-django
djangorestframework
djangorestframework-simplejwt